refactor: 승패 계산과 수익 계산 로직 이동
