+++ src/main/kotlin/blackjack/domain/participant/Player.kt
@@ -1,7 +1,5 @@
 package blackjack.domain.participant
 
-import blackjack.domain.GameResult
-import blackjack.domain.Money
 import blackjack.domain.ParticipantCards
 import blackjack.domain.ParticipantCards.Companion.BUST_STANDARD
 import blackjack.domain.card.TrumpCard
@@ -9,40 +7,12 @@ import blackjack.domain.card.TrumpCard
 class Player(
     val name: String,
     cards: ParticipantCards,
-    val bettingMoney: Money,
 ) : Participant(cards) {
     override fun showInitialCards(): List<TrumpCard> = takeCards(PLAYER_INITIAL_CARD_COUNT)
 
     override fun isDrawable(): Boolean = cards.sumOfCards <= BUST_STANDARD
 
-    override fun getResult(other: Participant): GameResult {
-        val myScore = cards.finalScore()
-        val otherScore = other.cards.finalScore()
-
-        return when {
-            other.cards.isBlackJack() && !cards.isBlackJack() -> GameResult.LOSE
-            cards.isBlackJack() && !other.cards.isBlackJack() -> GameResult.BLACKJACK
-            cards.isBust() -> GameResult.LOSE
-            other.cards.isBust() && !cards.isBust() -> GameResult.WIN
-            myScore > otherScore -> GameResult.WIN
-            myScore < otherScore -> GameResult.LOSE
-            else -> GameResult.DRAW
-        }
-    }
-
-    override fun getProfit(gameResult: GameResult): Double =
-        when (gameResult) {
-            GameResult.BLACKJACK -> bettingMoney.value * PLAYER_BLACKJACK_MULTIPLY
-            GameResult.WIN -> bettingMoney.value * WIN_MULTIPLY
-            GameResult.DRAW -> DRAW_MULTIPLY
-            else -> bettingMoney.value * LOSE_MULTIPLY
-        }
-
     companion object {
         private const val PLAYER_INITIAL_CARD_COUNT = 2
-        private const val PLAYER_BLACKJACK_MULTIPLY = 1.5
-        const val WIN_MULTIPLY = 1.0
-        const val DRAW_MULTIPLY = 0.0
-        const val LOSE_MULTIPLY = -1.0
     }
 }
